Add files via upload
There is considerable variability in the number of bikes available across different stations. The distribution of bikes is not uniform, indicating that bike availability is not evenly spread out across the stations. Some stations are consistently overserved or underserved. The stations with the highest peaks may indicate over-saturation of bikes, suggesting an opportunity to redistribute bikes more efficiently to meet demand across the network in the city.